requirements.txt
openai
langchain
langchain-openai
langgraph
chromadb
streamlit
fastapi
uvicorn
pandas
python-dotenv
